Разделил код на модули и импортировал их
